Moved Make_Resumed_File.py to deprecated
unnecesary now that have file maker in interface
